requirements.txt
requests>=2.28.0
pyyaml>=6.0
google-genai>=1.0.0
schedule>=1.2.0
